Updated PID in Auton

+++ HighStakes/Competition/moveBot/include/AutonFunctions.h
@@ -7,7 +7,7 @@
 #include "api.h"
 
 // Function declarations
-void moveForwardPID(int targetDistance, int maxSpeed);
+void moveForwardPID(double targetDistance, int maxSpeed);
 void turnClockwiseTime(int turnTime, int maxSpeed);
 double degreesToInches(double);
 // PID Constants (extern to be defined in AutonFunctions.cpp)
